feat: migrate to Bun, convert to ESM/TS, and add multi-platform build scripts for GitHub Releases
